+++ did-ethr/src/lib.rs
@@ -138,6 +138,13 @@ impl DIDMethod for DIDEthr {
     fn generate(&self, source: &Source) -> Option<String> {
         let jwk = match source {
             Source::Key(jwk) => jwk,
+            Source::KeyAndPattern(jwk, pattern) => {
+                if !pattern.is_empty() {
+                    // TODO: support pattern
+                    return None;
+                }
+                jwk
+            }
             _ => return None,
         };
         let hash = match ssi::keccak_hash::hash_public_key(jwk) {