+++ did-sol/src/lib.rs
@@ -140,6 +140,13 @@ impl DIDMethod for DIDSol {
     fn generate(&self, source: &Source) -> Option<String> {
         let jwk = match source {
             Source::Key(jwk) => jwk,
+            Source::KeyAndPattern(jwk, pattern) => {
+                if !pattern.is_empty() {
+                    // pattern not supported
+                    return None;
+                }
+                jwk
+            }
             _ => return None,
         };
         let did = match jwk.params {